Move folder export tests
